beginnings of player stat implementation
minor order change
