+++ src/app/page.tsx
@@ -1,6 +1,5 @@
 "use client";
-import Dps from "./dps";
-
+import Dps from "./Dps";
 export default function Home() {
   return <Dps></Dps>;
 }
